refactor: create Keyboard class/refactor code
