+++ codejam-virtual-keyboard/style/main.css
@@ -53,7 +53,7 @@ body {
   color: #000;
   box-shadow: 0 0 4px rgba(0, 0, 0, 0.4);
 
-  transition: background-color 0.1s, transform 0.01s, color 0.1s;
+  transition: background-color 0.1s, transform 0.1s, color 0.1s;
   cursor: pointer;
 }
 
@@ -93,6 +93,5 @@ body {
 .pressed-btn {
   color: #ffffff;
   background: #f9a825 !important;
-	transform: scale(0.9);
-	transform: skewX(5deg);
+	transform: scale(0.85) skewX(5deg);
 }
